requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
httpx==0.24.0
pydantic==1.10.7  # Añade esto si necesitas una versión específica
